requirement.txt
Django == 1.11
DjangoUeditor == 1.8.143
MySQL-python == 1.2.5
Pillow == 4.3.0
PyYAML == 3.12
asn1crypto == 0.23.0
cffi == 1.11.2
cryptography == 2.1.4
diff-match-patch == 20121119
django-crispy-forms == 1.7.0
django-formtools == 1
django-import-export == 0.6.0
django-pure-pagination == 0.3.0
django-ranged-response == 0.2.0
django-reversion == 2.0.11
django-simple-captcha == 0.5.6
enum34 == 1.1.6
et-xmlfile == 1.0.1
future == 0.16.0
httplib2 == 0.9.2
idna == 2.6
ipaddress == 1.0.18
jdcal == 1.3
lazy-property == 0.0.1
ndg-httpsclient == 0.4.3
odfpy == 1.3.5
olefile == 0.44
openpyxl == 2.4.9
pip == 9.0.1
pyOpenSSL == 17.5.0
pyasn1 == 0.4.2
pycparser == 2.18
pytz == 2017.3
requests == 2.5.3
setuptools == 38.2.3
six == 1.10.0
tablib == 0.12.1
unicodecsv == 0.14.1
wheel == 0.30.0
xlrd == 1.1.0
xlwt == 1.3.0

